fix(core): change TaskItem.Complexity from double to int
Complexity is only ever entered/validated as a whole number 1-10; narrow the type across the model, API DTOs, CLI add/edit handlers, and tests to match.

+++ PriorityTaskManager.API/Local/LocalScheduleDtos.cs
@@ -9,7 +9,7 @@ namespace PriorityTaskManager.API.Local
 		bool IsCompleted,
 		double Progress,
 		int Importance,
-		double Complexity,
+		int Complexity,
 		double Points,
 		DateTime? DueDate,
 		DateTime? NotBefore,
